requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==3.0.0
Django==4.2.16
django-allauth==0.52.0
django-axes==6.5.1
django-cors-headers==3.14.0
django-extensions==3.2.3
django-import-export==3.2.0
django-jazzmin==2.6.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-rest-captcha==0.3.0
django-simple-captcha==0.6.0
djangorestframework==3.14.0
dnspython==2.7.0
drf-yasg==1.21.5
elastic-transport==8.15.1
elasticsearch==8.16.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==1.26.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.2.0
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.17
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.4.4
starlette==0.41.2
tablib==3.5.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==14.1
xlrd==2.0.1
xlwt==1.3.0
